+++ udata/core/spatial/api.py
@@ -1,7 +1,7 @@
 from flask import current_app, abort
 from mongoengine.queryset.visitor import Q
 
-from flask_restplus import inputs
+from flask_restx import inputs
 
 from udata.api import api, API
 from udata.i18n import _