+++ udata/tests/apiv2/test_swagger.py
@@ -10,15 +10,15 @@ class SwaggerBlueprintTest:
     modules = []
 
     def test_swagger_resource_type(self, api):
-        response = api.get(url_for('api.specs'))
+        response = api.get(url_for('apiv2.specs'))
         assert200(response)
         swagger = json.loads(response.data)
         expected = swagger['paths']['/datasets/{dataset}/resources/']
-        expected = expected['put']['responses']['200']['schema']['type']
-        assert expected == 'array'
+        expected = expected['get']['responses']['200']['schema']['$ref']
+        assert expected == '#/definitions/ResourcePage'
 
     def test_swagger_specs_validate(self, api):
-        response = api.get(url_for('api.specs'))
+        response = api.get(url_for('apiv2.specs'))
         try:
             schemas.validate(response.json)
         except schemas.SchemaValidationError as e: